third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-ui/callstack-placards-discarded-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that RawSourceCode listeners count won't grow on each script pause. Bug 70996


Running: testCallStackPlacardsDiscarded
Script source was shown.
Set timer for test function.
Received DebuggerPaused event.
Function name: testFunction
Script execution paused.
Script execution resumed.
Set timer for test function.
Received DebuggerPaused event.
Function name: testFunction
Script execution paused.
Script execution resumed.
